put the various languages in packages to skip tests when a specific language is not relevant for the user
