+++ OLGA/OLGA-Core/src/test/java/semanticstore/ontology/library/generator/generators/csharp/PartialOptionTest.java
@@ -1,0 +1,68 @@
+/*
+ * -------------------------
+ * 
+ * MIT License
+ * 
+ * Copyright (c) 2018, Schneider Electric USA, Inc.
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * 
+ * ---------------------
+ */
+package semanticstore.ontology.library.generator.code.generator.generators.csharp;
+
+import static org.junit.Assert.assertTrue;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.stream.Stream;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import semanticstore.ontology.library.generator.test.utils.GeneratedOntologies;
+
+public class PartialOptionTest {
+
+  @BeforeClass
+  public static void generateOntology() throws Exception {
+    // check if the needed ontologies have been generated
+    assertTrue(GeneratedOntologies.isSimpleBasicPartialGenerated());
+  }
+
+  @Test
+  public void testPartial() throws IOException {
+    Path pathToClasses =
+        Paths.get("OLGA/generated/testSimplePartial-dotnetTrinity/TestSimplePartial/Rdf/Model/");
+    try (Stream<String> stream = Files.lines(pathToClasses.resolve("Building.cs"))) {
+      assertTrue(stream.anyMatch(line -> line.contains("partial")));
+    }
+    try (Stream<String> stream = Files.lines(pathToClasses.resolve("IBuilding.cs"))) {
+      assertTrue(stream.anyMatch(line -> line.contains("partial")));
+    }
+  }
+
+  @Test
+  public void testPartialNotSet() throws IOException {
+    Path pathToClasses = Paths.get("OLGA/generated/testSimple-dotnetTrinity/TestSimple/Rdf/Model/");
+    try (Stream<String> stream = Files.lines(pathToClasses.resolve("Building.cs"))) {
+      assertTrue(!stream.anyMatch(line -> line.contains("partial")));
+    }
+    try (Stream<String> stream = Files.lines(pathToClasses.resolve("IBuilding.cs"))) {
+      assertTrue(!stream.anyMatch(line -> line.contains("partial")));
+    }
+  }
+
+}
